nginx_conf.sh is changed-final

+++ nginx_conf.sh
@@ -1,28 +1,47 @@
 #!/bin/bash
 
-# Updating and installing required commands
-apt update  -y
+# installing required commands
+VeriableWget="/usr/bin/wget"
+if [[ ! -e $VeriableWget ]]
+then
 apt install wget -y
-apt install curl -y
-
-# initiating variables
-URL=https://transfer.sh/QX7UuX/bootcamp_aca.conf
-response=$(curl --write-out %{http_code} --silent --output /dev/null $URL)
-
-# Checking if the given URL works and if we already have one
-if [[ $response = 200 ]]; then
-	wget $URL
 fi
 
-# checking if we already have the copy as our nginx default conf
-deff=$(diff bootcamp_aca.conf /etc/nginx/sites-enabled/default)
+VeriableCurl="/usr/bin/curl"
+if [[ ! -e $VeriableCurl ]]
+then
+apt install curl -y
+fi
 
-if [[ $deff -ne 0 ]]; then 
-       cp bootcamp_aca.conf /etc/nginx/sites-enabled/default
+# initiating variables
+url='https://transfer.sh/QX7UuX/bootcamp_aca.conf'
+response=$(curl --write-out %{http_code} --silent --output /dev/null $url)
+file="bootcamp_aca.conf"
+# Checking if the given URL works and if we already have one
+if [[ $response = 200 ]]; then
+	wget $url
+fi
+
+# chechking if file is zero or not 
+if [[ ! -z $file ]]; then 
+       mv bootcamp_aca.conf /etc/nginx/sites-enabled/
+fi
+
+#remove /etc/nginx/sites-enabled/default file
+
+if [[ -f /etc/nginx/sites-enable/default ]]
+then
+rm /etc/nginx/sites-enabled/default
+
+fi
+
+#testing nginx
+nginx -t
+Vernginx=$(echo $?)
+if [[ ! $Vernginx == 0 ]]; then
+	exit
 fi
 
 # reloading service
 service nginx reload
 
-# removing the conf file
-rm -f bootcamp_aca.conf
